VESCのRPM入出力をInt16に統一

+++ ros2_ws/src/robot_controller/belt_controller/include/belt_controller.hpp
@@ -1,19 +1,19 @@
 #pragma once
 
 #include <cstdint>
+#include <limits>
 #include <string>
 
 #include "rclcpp/rclcpp.hpp"
 #include "std_msgs/msg/bool.hpp"
-#include "std_msgs/msg/float32.hpp"
+#include "std_msgs/msg/int16.hpp"
 #include "std_msgs/msg/u_int8.hpp"
 
-class BeltControllerNode : public rclcpp::Node
-{
-public:
+class BeltControllerNode : public rclcpp::Node {
+ public:
   BeltControllerNode();
 
-private:
+ private:
   static constexpr uint8_t stop_mode_{1};
   static constexpr uint8_t level_1_mode_{2};
   static constexpr uint8_t level_2_mode_{3};
@@ -22,26 +22,22 @@ private:
   void declare_parameters();
   void get_parameters();
   void belt_mode_callback(const std_msgs::msg::UInt8::SharedPtr msg);
-  void underbelt_feedback_callback(
-    const std_msgs::msg::Float32::SharedPtr msg);
-  void upperbelt_feedback_callback(
-    const std_msgs::msg::Float32::SharedPtr msg);
+  void underbelt_feedback_callback(const std_msgs::msg::Int16::SharedPtr msg);
+  void upperbelt_feedback_callback(const std_msgs::msg::Int16::SharedPtr msg);
   void timer_callback();
-  double target_rpm_from_mode(uint8_t mode);
-  bool is_rpm_valid(double rpm) const;
-  bool is_belt_ready(
-    double underbelt_target_rpm,
-    double upperbelt_target_rpm,
-    const rclcpp::Time & current_time);
+  int target_rpm_from_mode(uint8_t mode);
+  bool is_rpm_valid(int rpm) const;
+  bool is_belt_ready(int underbelt_target_rpm, int upperbelt_target_rpm,
+                     const rclcpp::Time& current_time);
 
   bool is_configuration_valid_{true};
   uint8_t belt_mode_{stop_mode_};
-  double stop_rpm_{0.0};
-  double level_1_rpm_{1000.0};
-  double level_2_rpm_{2000.0};
-  double level_3_rpm_{3000.0};
+  int stop_rpm_{0};
+  int level_1_rpm_{1000};
+  int level_2_rpm_{2000};
+  int level_3_rpm_{3000};
   int command_period_ms_{10};
-  double ready_tolerance_rpm_{100.0};
+  int ready_tolerance_rpm_{100};
   double ready_hold_sec_{0.1};
   int qos_depth_{1};
   std::string belt_mode_topic_;
@@ -50,19 +46,17 @@ private:
   std::string underbelt_current_rpm_topic_;
   std::string upperbelt_current_rpm_topic_;
   std::string belt_ready_topic_;
-  double underbelt_current_rpm_{0.0};
-  double upperbelt_current_rpm_{0.0};
+  int underbelt_current_rpm_{0};
+  int upperbelt_current_rpm_{0};
   bool underbelt_feedback_received_{false};
   bool upperbelt_feedback_received_{false};
   rclcpp::Time ready_since_{};
 
   rclcpp::Subscription<std_msgs::msg::UInt8>::SharedPtr belt_mode_sub_;
-  rclcpp::Subscription<std_msgs::msg::Float32>::SharedPtr
-    underbelt_feedback_sub_;
-  rclcpp::Subscription<std_msgs::msg::Float32>::SharedPtr
-    upperbelt_feedback_sub_;
-  rclcpp::Publisher<std_msgs::msg::Float32>::SharedPtr underbelt_rpm_pub_;
-  rclcpp::Publisher<std_msgs::msg::Float32>::SharedPtr upperbelt_rpm_pub_;
+  rclcpp::Subscription<std_msgs::msg::Int16>::SharedPtr underbelt_feedback_sub_;
+  rclcpp::Subscription<std_msgs::msg::Int16>::SharedPtr upperbelt_feedback_sub_;
+  rclcpp::Publisher<std_msgs::msg::Int16>::SharedPtr underbelt_rpm_pub_;
+  rclcpp::Publisher<std_msgs::msg::Int16>::SharedPtr upperbelt_rpm_pub_;
   rclcpp::Publisher<std_msgs::msg::Bool>::SharedPtr belt_ready_pub_;
   rclcpp::TimerBase::SharedPtr timer_;
 };
